+++ src/test/java/domain/card/DeckTest.java
@@ -1,6 +1,5 @@
-package domain;
+package domain.card;
 
-import domain.card.*;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
