+++ src/test/java/domain/card/HandTest.java
@@ -1,9 +1,5 @@
-package domain;
+package domain.card;
 
-import domain.card.Card;
-import domain.card.Hand;
-import domain.card.Rank;
-import domain.card.Suit;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
