+++ src/test/java/domain/participant/BetAmountTest.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.participant;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
