+++ src/test/java/domain/participant/DealerTest.java
@@ -1,9 +1,8 @@
-package domain;
+package domain.participant;
 
 import domain.card.Card;
 import domain.card.Rank;
 import domain.card.Suit;
-import domain.participant.Dealer;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
