+++ src/test/java/domain/participant/PlayerTest.java
@@ -1,11 +1,8 @@
-package domain;
+package domain.participant;
 
 import domain.card.Card;
 import domain.card.Rank;
 import domain.card.Suit;
-import domain.participant.Name;
-import domain.participant.Participant;
-import domain.participant.Player;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
