+++ src/test/java/domain/result/ResultTest.java
@@ -1,6 +1,5 @@
-package domain;
+package domain.result;
 
-import domain.result.Result;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
